dp: pt: rearrange folder structure of atomic_model and model.
