+++ deepmd/dpmodel/model/__init__.py
@@ -1,16 +1,8 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
-from .dp_atomic_model import (
-    DPAtomicModel,
-)
 from .dp_model import (
     DPModel,
-)
-from .make_base_atomic_model import (
-    make_base_atomic_model,
 )
 
 __all__ = [
     "DPModel",
-    "DPAtomicModel",
-    "make_base_atomic_model",
 ]
