+++ deepmd/pt/model/atomic_model/base_atomic_model.py
@@ -2,7 +2,7 @@
 
 import torch
 
-from deepmd.dpmodel.model import (
+from deepmd.dpmodel.atomic_model import (
     make_base_atomic_model,
 )
