+++ deepmd/pt/model/model/__init__.py
@@ -1,22 +1,25 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
 import copy
 
+from deepmd.pt.model.atomic_model import (
+    DPAtomicModel,
+    PairTabAtomicModel,
+)
 from deepmd.pt.model.descriptor.descriptor import (
     Descriptor,
-)
-from deepmd.pt.model.model.dp_atomic_model import (
-    DPAtomicModel,
-)
-from deepmd.pt.model.model.pairtab_atomic_model import (
-    PairTabModel,
 )
 from deepmd.pt.model.task import (
     Fitting,
 )
 
-from .ener import (
+from .dp_model import (
+    DPModel,
+)
+from .dp_zbl_model import (
+    DPZBLModel,
+)
+from .ener_model import (
     EnergyModel,
-    ZBLModel,
 )
 from .model import (
     BaseModel,
@@ -44,13 +47,13 @@ def get_zbl_model(model_params):
     dp_model = DPAtomicModel(descriptor, fitting, type_map=model_params["type_map"])
     # pairtab
     filepath = model_params["use_srtab"]
-    pt_model = PairTabModel(
+    pt_model = PairTabAtomicModel(
         filepath, model_params["descriptor"]["rcut"], model_params["descriptor"]["sel"]
     )
 
     rmin = model_params["sw_rmin"]
     rmax = model_params["sw_rmax"]
-    return ZBLModel(
+    return DPZBLModel(
         dp_model,
         pt_model,
         rmin,
@@ -84,4 +87,5 @@ __all__ = [
     "BaseModel",
     "EnergyModel",
     "get_model",
+    "DPModel",
 ]
