+++ deepmd/pt/model/model/dp_model.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
-from deepmd.dpmodel import (
+from deepmd.pt.model.atomic_model import (
     DPAtomicModel,
 )
