+++ deepmd/pt/model/model/ener_model.py
@@ -6,87 +6,9 @@ from typing import (
 
 import torch
 
-from .dp_atomic_model import (
-    DPAtomicModel,
+from .dp_model import (
+    DPModel,
 )
-from .linear_atomic_model import (
-    DPZBLLinearAtomicModel,
-)
-from .make_model import (
-    make_model,
-)
-
-DPModel = make_model(DPAtomicModel)
-ZBLModel_ = make_model(DPZBLLinearAtomicModel)
-
-
-class ZBLModel(ZBLModel_):
-    model_type = "ener"
-
-    def __init__(
-        self,
-        *args,
-        **kwargs,
-    ):
-        super().__init__(*args, **kwargs)
-
-    def forward(
-        self,
-        coord,
-        atype,
-        box: Optional[torch.Tensor] = None,
-        fparam: Optional[torch.Tensor] = None,
-        aparam: Optional[torch.Tensor] = None,
-        do_atomic_virial: bool = False,
-    ) -> Dict[str, torch.Tensor]:
-        model_ret = self.forward_common(
-            coord,
-            atype,
-            box,
-        )
-
-        model_predict = {}
-        model_predict["atom_energy"] = model_ret["energy"]
-        model_predict["energy"] = model_ret["energy_redu"]
-        if self.do_grad("energy"):
-            model_predict["force"] = model_ret["energy_derv_r"].squeeze(-2)
-            if do_atomic_virial:
-                model_predict["atom_virial"] = model_ret["energy_derv_c"].squeeze(-3)
-            model_predict["virial"] = model_ret["energy_derv_c_redu"].squeeze(-2)
-        else:
-            model_predict["force"] = model_ret["dforce"]
-        return model_predict
-
-    def forward_lower(
-        self,
-        extended_coord,
-        extended_atype,
-        nlist,
-        mapping: Optional[torch.Tensor] = None,
-        do_atomic_virial: bool = False,
-    ):
-        model_ret = self.forward_common_lower(
-            extended_coord,
-            extended_atype,
-            nlist,
-            mapping,
-        )
-
-        model_predict = {}
-        model_predict["atom_energy"] = model_ret["energy"]
-        model_predict["energy"] = model_ret["energy_redu"]
-        if self.do_grad("energy"):
-            model_predict["extended_force"] = model_ret["energy_derv_r"].squeeze(-2)
-            model_predict["virial"] = model_ret["energy_derv_c_redu"].squeeze(-2)
-            if do_atomic_virial:
-                model_predict["extended_virial"] = model_ret["energy_derv_c"].squeeze(
-                    -2
-                )
-        else:
-            assert model_ret["dforce"] is not None
-            model_predict["dforce"] = model_ret["dforce"]
-        model_predict = model_ret
-        return model_predict
 
 
 class EnergyModel(DPModel):
@@ -109,7 +31,12 @@ class EnergyModel(DPModel):
         do_atomic_virial: bool = False,
     ) -> Dict[str, torch.Tensor]:
         model_ret = self.forward_common(
-            coord, atype, box, do_atomic_virial=do_atomic_virial
+            coord,
+            atype,
+            box,
+            fparam=fparam,
+            aparam=aparam,
+            do_atomic_virial=do_atomic_virial,
         )
         if self.fitting_net is not None:
             model_predict = {}
@@ -135,6 +62,8 @@ class EnergyModel(DPModel):
         extended_atype,
         nlist,
         mapping: Optional[torch.Tensor] = None,
+        fparam: Optional[torch.Tensor] = None,
+        aparam: Optional[torch.Tensor] = None,
         do_atomic_virial: bool = False,
     ):
         model_ret = self.forward_common_lower(
@@ -142,6 +71,8 @@ class EnergyModel(DPModel):
             extended_atype,
             nlist,
             mapping,
+            fparam=fparam,
+            aparam=aparam,
             do_atomic_virial=do_atomic_virial,
         )
         if self.fitting_net is not None:
